+++ py_partiql_parser/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.5.0"
+__version__ = "0.5.1"
 
 
 from ._internal.parser import DynamoDBStatementParser, S3SelectParser  # noqa